+++ sqlite3_web/test/client_test.dart
@@ -15,7 +15,7 @@ void main() {
     expect(all, [
       DatabaseImplementation.opfsShared,
       DatabaseImplementation.opfsWithExternalLocks,
-      DatabaseImplementation.opfsAtomics,
+      DatabaseImplementation.opfsWithExternalLocksWorkaround,
       DatabaseImplementation.indexedDbShared,
       DatabaseImplementation.indexedDbUnsafeWorker,
       DatabaseImplementation.indexedDbUnsafeLocal,